notes to self
